goa_voter_card_scrappy/requirements.txt
apturl==0.5.2
attrs==20.3.0
Automat==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
flasgger==0.9.5
Flask==1.1.2
future==0.18.2
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
imagetyperzapi3==0.5
importlib-metadata==3.7.3
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mistune==0.8.4
mock==3.0.5
monotonic==1.5
mysql-connector-python==8.0.15
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parsel==1.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
Protego==0.1.16
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyDispatcher==2.0.5
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==20.0.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
QtPy==1.9.0
queuelib==1.5.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
Scrapy==2.4.1
scrapy-mysql-pipeline==2019.7.19
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==2.2
systemd-python==234
Twisted==21.2.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
w3lib==1.22.0
wadllib==1.3.3
Werkzeug==1.0.1
xkit==0.0.0
zipp==3.4.1
zope.interface==5.2.0
